LayoutTests/imported/w3c/web-platform-tests/css/cssom/stylesheet-same-origin.sub-expected.txt


FAIL Origin-clean check in cross-origin CSSOM Stylesheets assert_throws_dom: stylesheet.insertRule should throw SecurityError. function "function () {
                    sheet.insertRule("#test { margin: 10px; }", 1);
                }" did not throw
FAIL Origin-clean check in cross-origin CSSOM Stylesheets (redirect from same-origin to cross-origin) assert_throws_dom: stylesheet.insertRule should throw SecurityError. function "function () {
                    sheet.insertRule("#test { margin: 10px; }", 1);
                }" did not throw
FAIL Origin-clean check in cross-origin CSSOM Stylesheets (redirect from cross-origin to same-origin) assert_throws_dom: stylesheet.cssRules should throw SecurityError. function "function () {
                    sheet.cssRules;
                }" did not throw
FAIL Origin-clean check in loading error CSSOM Stylesheets assert_throws_dom: stylesheet.cssRules should throw SecurityError. function "function () {
                    sheet.cssRules;
                }" did not throw
PASS Origin-clean check in same-origin CSSOM Stylesheets
FAIL Origin-clean check in data:css CSSOM Stylesheets Not allowed to access cross-origin stylesheet
